Collapsed linode_api.object.profile into one file
