--- linode_api/objects/profile/__init__.py
@@ -1,4 +1,0 @@
-from .personal_access_token import PersonalAccessToken
-from .app import AuthorizedApp
-from .profile import Profile
-from .whitelist_entry import WhitelistEntry
