--- linode_api/objects/profile/app.py
@@ -1,15 +1,0 @@
-from linode_api.objects import Base, Property
-
-
-class AuthorizedApp(Base):
-    api_endpoint = "/profile/apps/{id}"
-
-    properties = {
-        "id": Property(identifier=True),
-        "scopes": Property(),
-        "label": Property(),
-        "created": Property(is_datetime=True),
-        "expiry": Property(is_datetime=True),
-        "thumbnail_url": Property(),
-        "website": Property(),
-    }
