--- linode_api/objects/profile/personal_access_token.py
@@ -1,16 +1,0 @@
-from __future__ import absolute_import
-
-from linode_api.objects import Base, Property
-
-
-class PersonalAccessToken(Base):
-    api_endpoint = "/profile/tokens/{id}"
-
-    properties = {
-        "id": Property(identifier=True),
-        "scopes": Property(),
-        "label": Property(mutable=True),
-        "created": Property(is_datetime=True),
-        "token": Property(),
-        "expiry": Property(is_datetime=True),
-    }
