--- linode_api/objects/profile/whitelist_entry.py
@@ -1,14 +1,0 @@
-from __future__ import absolute_import
-
-from linode_api.objects import Base, Property
-
-
-class WhitelistEntry(Base):
-    api_endpoint = "/profile/whitelist/{id}"
-
-    properties = {
-        'id': Property(identifier=True),
-        'address': Property(),
-        'netmask': Property(),
-        'note': Property(),
-    }
